python3: added compatibility
Dodgy solution in which we create a python2 and
python3 compatible wrapper that would import the
old script, or a new automatically migrated
version.

The automatic migration was done with the 2to3
tool after modifying the original script so we can
invoke it as a module.

This is needed since LibreOffice is providing its
own python3 bundled package since 4.x series.