Handle exit code 81 and reconnect if encountered.
If user installation files do not already exist, the first time LibreOffice starts it will create the files and then immediately quit with exit code 81 (I'm not sure about OpenOffice, but I assume it works similarly). In this case, unoconv will fail, because the listener will not remain open long enough to be used. This behaviour will happen repeatedly if the env:UserInstallation variable is set when launching soffice.

I have tested out this scenario in two environments (Mac OSX 10.9 and Ubuntu 12.04). On the mac, unoconv will launch the executable soffice, whereas on Ubuntu, it launches soffice.bin (this file is not present on MacOSX, at least not with LO 4.2). As far as I can tell, the difference between soffice and soffice.bin is that soffice is a wrapper around soffice.bin (where available), and includes logic to automatically restart the executable when an exit code 81 is encountered. Therefore, on MacOSX, I do not run into the problem of unoconv failing because of exit code 81, but on ubuntu I do.

As mentioned, soffice (no .bin) is present on Ubuntu as well (again, at least for LO 4.2), and using that executable instead of soffice.bin seems to have no adverse effects. Does anyone know why the choice was made to use soffice.bin instead?