Allow using a custom user profile (#451)
This is useful when converting large documents on a machine with many
CPU cores/threads. Using a single unoconv server means in fact a single
document is converted at a time, given LibreOffice uses a single global
lock for remote method invocations.

Starting multiple unoconv servers is not possible by default, as the
second one will complain that the user profile is locked. Using a custom
user profile is a solution to this.