Detect an unusable Home directory
When Open/LibreOffice runs, it looks for a folder containing user-specific
configuration in the invoking user's home directory. If it cannot find it,
Open/LibreOffice attempts to create it.

For users running `unoconv` manually, this isn't a problem.

For users attempting to run this on a webserver, the invoking user is the http
daemon (e.g. `apache`, `nginx`, `lighttpd`), and the home folder is typically
the webserver's root directory (e.g. `/srv/http`, `/var/www`).

Unfortunately, there appear to be some scenarios where this folder is read-only,
even for the daemon (see #87, #449). This leads to errors, as Open/LibreOffice
is unable to create its configuration folder.

(It is assumed that if the configuration folder already exists, it must be
usable, else how would it have been created?)

This commit solves the issue by detecting this, emitting a warning message, and
using a temporary folder as a temporary "Home".